refactor(tangle-dapp, icons): Improve inconsistent naming
