--- apps/tangle-dapp/containers/DelegationsPayoutsContainer/DelegatorTableContainer.tsx
@@ -1,12 +1,0 @@
-import { DelegatorTable } from '../../components';
-import { Delegator } from '../../types';
-
-export default function DelegatorTableContainer({
-  pageSize,
-  value,
-}: {
-  pageSize: number;
-  value: Delegator[];
-}) {
-  return <DelegatorTable data={value} pageSize={pageSize} />;
-}
