--- apps/tangle-dapp/containers/DelegationsPayoutsContainer/index.ts
@@ -1,1 +1,0 @@
-export { default as DelegationsPayoutsContainer } from './DelegationsPayoutsContainer';
