+++ apps/tangle-dapp/containers/NominationsPayoutsContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as NominationsPayoutsContainer } from './NominationsPayoutsContainer';
