+++ apps/tangle-dapp/containers/index.ts
@@ -1,9 +1,9 @@
 export * from './BondMoreTxContainer';
 export * from './DelegateTxContainer';
-export * from './DelegationsPayoutsContainer';
 export * from './HeaderChipsContainer';
 export * from './KeyStatsContainer';
 export { Layout } from './Layout';
+export * from './NominationsPayoutsContainer';
 export * from './NominatorStatsContainer';
 export * from './RebondTxContainer';
 export * from './TxConfirmationModalContainer';
